requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.3
undetected-chromedriver==3.5.5
google-genai==0.3.0